My_first_Agent/requirements.txt
fastapi
uvicorn
openai
pandas
duckdb
pydantic
python-dotenv
